requirements.txt
aiohttp==3.8.4
aiohttp-session==2.12.0
aiopg==1.4.0
alembic==1.10.4
asyncpg==0.27.0
SQLAlchemy
PyYAML